feat: optimize hooks
